Backfill specs
This required sinning. The validation code that gets generated makes a mess
out of reading pipeline YAML, so I had to take it out manually. I also had
to tear out the generated placeholder controller test.

The nice bit in this commit is the splitting of the main e2e spec into smaller
bits.